src/main/java/com/example/recipe/converters/CategoryToCategoryCommand.java
package com.example.recipe.converters;

import com.example.recipe.commands.CategoryCommand;
import com.example.recipe.domain.Category;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

@Component
public class CategoryToCategoryCommand implements Converter<Category, CategoryCommand> {

    @Synchronized
    @Nullable
    @Override
    public CategoryCommand convert(Category category) {
        if (category != null) {
            final CategoryCommand command = new CategoryCommand();
            command.setId(category.getId());
            command.setName(category.getName());
            return command;
        }
        return null;
    }
}
